Restructured files and fixed eslint warnings
