+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-task-tracker",
+  "name": "a-finman",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
