+++ src/components/home/HomePageButton.js
@@ -1,10 +1,10 @@
-import "./HomePageButton.css";
+import "../../styles/HomePageButton.css";
 
 const HomePageButton = ({ text1, text2, onClick }) => {
   return (
     <button onClick={onClick} className="homepagebutton">
-      <a className="bigpart">{text1}</a>
-      <a className="smallpart">{text2}</a>
+      <span className="bigpart">{text1}</span>
+      <span className="smallpart">{text2}</span>
     </button>
   );
 };
