+++ src/components/pdf/table-items/InvoiceTableBlankSpace.js
