--- src/components/seeInvoicesComponents/pdfStyling.css
@@ -1,8 +1,0 @@
-.customOverlay {
-}
-
-.customModal {
-  width: 1000px;
-  padding: 50px;
-  margin: 50px;
-}
